+++ components/dashboard/PasswordManager.cy.tsx
@@ -2,14 +2,34 @@
 
 import { Inter } from 'next/font/google';
 import React, { CSSProperties } from 'react';
-import { User, UserAction } from 'contexts/reducers/user-reducer';
-import { UserProvider, useUser, useUserDispatch } from 'contexts/use-user';
+import { User, UserAction, UserStatus } from 'contexts/reducers/user.reducer';
+import { UserProvider, useUser, useUserDispatch } from 'contexts/user.context';
 import { IMAGE_FILE } from 'lib/images';
 import PasswordManager from './PasswordManager';
-import { TagEntriesProvider } from 'contexts/use-tag-entries';
-import { PasswordEntriesProvider } from 'contexts/use-password-entries';
-
+import { TagEntriesProvider } from 'contexts/tag-entries.context';
+import { PasswordEntriesProvider } from 'contexts/password-entries.context';
+import { Dropbox, DropboxAuth, DropboxResponse, files } from 'dropbox';
+import { LocationProvider } from 'contexts/location.context';
 const inter = Inter({ subsets: ['latin'] });
+
+const LOGGED_IN_USER: User = {
+  status: UserStatus.ONLINE_WITH_TREZOR,
+  dbc: getDbxMock(),
+  device:{
+    label: 'trezor_device_label',
+    appDataSeed: 'appDataSeed',
+    appDataEncryptionKey: new Uint8Array(32),
+    deviceId: 'deviceId',
+    model: 't1',
+    path: 'path',
+  },
+  dropboxAccountName: 'test',
+}
+
+function getDbxMock() {
+
+  return new Dropbox({ auth: new DropboxAuth({ clientId: '123' }) });
+}
 
 interface DashboardPageProps {
   children: React.ReactNode;
@@ -19,15 +39,17 @@ interface DashboardPageProps {
 function DashboardPageWrapper(props: DashboardPageProps) {
   return (
     <div className={inter.className} style={{margin: 0}}>
-      <UserProvider initialUser={props.initialUser}>
-        <TagEntriesProvider>
-          <PasswordEntriesProvider>
-            <DashboardPageController {...props}>
-              {props.children}
-            </DashboardPageController>
-          </PasswordEntriesProvider>
-        </TagEntriesProvider>
-      </UserProvider>
+      <LocationProvider>
+        <UserProvider initialUser={props.initialUser}>
+          <TagEntriesProvider>
+            <PasswordEntriesProvider>
+              <DashboardPageController {...props}>
+                {props.children}
+              </DashboardPageController>
+            </PasswordEntriesProvider>
+          </TagEntriesProvider>
+        </UserProvider>
+      </LocationProvider>
     </div>
   )
 }
@@ -49,7 +71,7 @@ function DashboardPageController({children, initialUser, onStorageLogin}: Dashbo
   )
 }
 
-describe('<Home />', () => {
+describe('Password Manager Page Tests', () => {
   beforeEach(() => {
     IMAGE_FILE.getPaths().forEach((image) => {
       cy.readFile(`Public${image}`, null).then((img) => {
@@ -63,12 +85,16 @@ describe('<Home />', () => {
     });
   });
 
-  it('renders password manager main page', () => {
+  it('password manager password interactions', () => {
+    const user = LOGGED_IN_USER;
     cy.viewport(1920,1080);
     cy.mount(
-      <DashboardPageWrapper>
+      <DashboardPageWrapper initialUser={user}>
         <PasswordManager />
       </DashboardPageWrapper>
-    );
+    ).then(() => {
+      debugger;
+    });
+
   });
 });
