+++ components/index/Home.cy.tsx
@@ -2,10 +2,10 @@
 import { Inter } from "next/font/google";
 import React, { CSSProperties} from 'react';
 import Home from './Home'
-import { User, UserProvider, useUser, useUserDispatch } from 'contexts/use-user';
+import { User, UserProvider, useUser, useUserDispatch } from 'contexts/user.context';
 import { IMAGE_FILE } from 'lib/images';
 import { Dropbox } from 'dropbox';
-import { UserAction } from '../../contexts/reducers/user-reducer';
+import { UserAction } from '../../contexts/reducers/user.reducer';
 
 const inter = Inter({ subsets: ['latin'] });
