+++ components/index/Home.tsx
@@ -4,10 +4,10 @@ import PinDialog from 'components/ui/PinDialog';
 import Image from 'next/image';
 import { IMAGE_FILE } from '../../lib/images';
 import styles from './Home.module.scss';
-import { UserStatus } from '../../contexts/reducers/user-reducer';
+import { UserStatus } from '../../contexts/reducers/user.reducer';
 import DeviceIcon from '../svg/ui/DeviceIcon';
 import Colors from 'styles/colors.module.scss';
-import { useUser } from '../../contexts/use-user';
+import { useUser } from '../../contexts/user.context';
 
 const LOGOUT_URL = 'https://www.dropbox.com/logout';
