+++ contexts/reducers/user.reducer.test.tsx
@@ -1,5 +1,5 @@
 import { describe, expect } from '@jest/globals';
-import { ActivatedTmpOnDevice, AddDevice, DropboxUserLoggedIn, User, userReducer, UserStatus } from './user-reducer';
+import { ActivatedTmpOnDevice, AddDevice, DropboxUserLoggedIn, User, userReducer, UserStatus } from './user.reducer';
 import { Dropbox } from 'dropbox';
 import { TrezorDevice } from '../../lib/trezor';
