+++ contexts/tag-entries.context.tsx
@@ -1,5 +1,5 @@
 import React, { createContext, Dispatch, useReducer } from 'react';
-import { TagEntries, TagsAction, tagsReducer, TagsStatus } from './reducers/tag-entries-reducer';
+import { TagEntries, TagsAction, tagsReducer, TagsStatus } from './reducers/tag-entries.reducer';
 import { TAG_ALL, TAG_BITCOIN, TAG_SOCIAL } from '../lib/images';
 
 const ALL_ID = '0';
