+++ contexts/user.context.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, Dispatch, MutableRefObject, useContext, useEffect, useReducer, useRef } from 'react';
-import { User, UserAction, userReducer, UserStatus } from './reducers/user-reducer';
+import { User, UserAction, userReducer, UserStatus } from './reducers/user.reducer';
 import { Dropbox } from 'dropbox';
-export { type User } from './reducers/user-reducer';
+export { type User } from './reducers/user.reducer';
 
 const defaultInitialUser = {
   status: UserStatus.OFFLINE,
