+++ lib/storage.ts
@@ -1,6 +1,6 @@
-import {} from '../contexts/use-password-entries';
-import { getSafePasswordEntries, PasswordEntries } from '../contexts/reducers/password-entries-reducer';
-import { getTags, TagEntry, TagEntries } from '../contexts/reducers/tag-entries-reducer';
+import {} from '../contexts/password-entries.context';
+import { getSafePasswordEntries, PasswordEntries } from '../contexts/reducers/password-entries.reducer';
+import { getTags, TagEntry, TagEntries } from '../contexts/reducers/tag-entries.reducer';
 import { SafePasswordEntry } from './trezor';
 import { uniqueId } from './utils';
 import entry from 'next/dist/server/typescript/rules/entry';
