+++ pages/dashboard.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
-import { useUser } from 'contexts/use-user';
-import { UserStatus } from 'contexts/reducers/user-reducer';
+import { useUser } from 'contexts/user.context';
+import { UserStatus } from 'contexts/reducers/user.reducer';
 import { useRouter } from 'next/router';
 import PasswordManager from 'components/dashboard/PasswordManager';
-import { Routes, useLocation } from '../contexts/use-location';
+import { Routes, useLocation } from '../contexts/location.context';
 
 export default function Dashboard() {
   const [location] = useLocation();
